Moving setup to setup
Now do all the setup steps in a specific setup file
